+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Add `always` function
 * Add `cycle` function
 * Add `sequential` function
+* Add `startWith` function
 
 ## 1.4.0 (2019-01-08)